requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
databases==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
mariadb==1.1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.45.3
stripe==10.12.0
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
